backend/requirements.txt
fastapi==0.111.1
sqlmodel==0.0.33
uvicorn[standard]==0.30.6
pyjwt==2.9.0
python-dotenv==1.0.1
pydantic>=2.7.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
asyncpg>=0.29.0
httpx>=0.27.0
psycopg2-binary>=2.9.10
bcrypt==4.0.1

# Phase III: Google Gemini (FREE with native tool calling)
google-generativeai>=0.8.0

# OpenAI (deprecated - quota exhausted)
# openai>=1.50.0

# Cohere (deprecated - no tool calling in free tier)
# cohere>=5.0.0

# MCP (kept for tool definitions)
mcp>=0.1.0

# Cohere (deprecated)
# cohere>=5.0.0
